Ver.1

+++ app/Http/Requests/StorePurchaseRequest.php
@@ -10,20 +10,20 @@ use Illuminate\Validation\Rule;
 
 class StorePurchaseRequest extends FormRequest
 {
-  public function authorize(): bool
-  {
-    return $this->user()?->can('create', Purchase::class) ?? false;
-  }
+    public function authorize(): bool
+    {
+        return $this->user()?->can('create', Purchase::class) ?? false;
+    }
 
-  public function rules(): array
-  {
-    return [
-      'supplier_id' => ['required', Rule::exists(Supplier::class, 'id')->where('is_active', true)],
-      'purchase_date' => ['required', 'date'],
-      'items' => ['required', 'array', 'min:1'],
-      'items.*.product_id' => ['required', 'distinct', Rule::exists(Product::class, 'id')->where('is_active', true)],
-      'items.*.quantity' => ['required', 'integer', 'min:1'],
-      'items.*.unit_price' => ['required', 'numeric', 'min:0'],
-    ];
-  }
+    public function rules(): array
+    {
+        return [
+            'supplier_id' => ['required', Rule::exists(Supplier::class, 'id')->where('is_active', true)],
+            'purchase_date' => ['required', 'date'],
+            'items' => ['required', 'array', 'min:1'],
+            'items.*.product_id' => ['required', 'distinct', Rule::exists(Product::class, 'id')->where('is_active', true)],
+            'items.*.quantity' => ['required', 'integer', 'min:1'],
+            'items.*.unit_price' => ['required', 'integer', 'min:0'],
+        ];
+    }
 }
